+++ src/main/java/virtuoel/pehkui/util/PehkuiEntitySelectorReaderExtensions.java
@@ -9,11 +9,11 @@ public interface PehkuiEntitySelectorReaderExtensions
 	void pehkui_setScaleType(final ScaleType scaleType);
 	
 	NumberRange<? extends Number> pehkui_getScaleRange();
-	void pehkui_setScaleRange(final NumberRange.FloatRange baseScaleRange);
+	void pehkui_setScaleRange(final NumberRange.DoubleRange baseScaleRange);
 	
 	ScaleType pehkui_getComputedScaleType();
 	void pehkui_setComputedScaleType(final ScaleType computedScaleType);
 	
 	NumberRange<? extends Number> pehkui_getComputedScaleRange();
-	void pehkui_setComputedScaleRange(final NumberRange.FloatRange computedScaleRange);
+	void pehkui_setComputedScaleRange(final NumberRange.DoubleRange computedScaleRange);
 }
